requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
Django==4.2.16
django-shell-plus==1.1.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
numpy==1.24.4
pandas==2.0.3
PyJWT==2.9.0
python-dateutil==2.9.0.post0
pytz==2024.2
six==1.16.0
sqlparse==0.5.1
typing-extensions==4.12.2
tzdata==2024.1
